# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillSync AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List available assessments
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Assessments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create an assessment (admin/recruiter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssessmentBody"
      responses:
        "201":
          description: Assessment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get assessment with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentDetail"
  /assessments/{id}/submit:
    post:
      operationId: submitAssessment
      tags: [assessments]
      summary: Submit assessment answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAssessmentBody"
      responses:
        "200":
          description: Assessment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
...
  /college/forms:
    get:
      operationId: listCollegeForms
      tags: [college]
      summary: List college forms
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Forms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollegeForm"
    post:
      operationId: createCollegeForm
      tags: [college]
      summary: Create a college form (faculty/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollegeFormBody"
      responses:
        "201":
          description: Form created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollegeForm"
  /college/forms/{id}:
    get:
      operationId: getCollegeForm
      tags: [college]
      summary: Get form with fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Form detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollegeForm"
  /college/submissions:
    get:
      operationId: listSubmissions
      tags: [college]
      summary: List form submissions
      parameters:
        - name: formId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormSubmission"
    post:
      operationId: createSubmission
      tags: [college]
      summary: Submit a college form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionBody"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormSubmission"
  /college/submissions/{id}:
    patch:
      operationId: updateSubmission
      tags: [college]
      summary: Approve or reject a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubmissionBody"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormSubmission"
  /college/announcements:
    get:
      operationId: listAnnouncements
      tags: [college]
      summary: List announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [college]
      summary: Create announcement (faculty/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  # NOTIFICATIONS
...
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, recruiter, faculty, freelancer_client, admin]
        institution:
          type: string
      required: [name, email, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    # USER
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, recruiter, faculty, freelancer_client, admin]
        avatar:
          type: string
        bio:
          type: string
        institution:
          type: string
        location:
          type: string
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        createdAt:
          type: string
      required: [id, name, email, role]
...
    Assessment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        type:
          type: string
          enum: [mcq, coding, aptitude, communication]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        duration:
          type: integer
        questionCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, category, type, difficulty, duration, questionCount]
    AssessmentDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        type:
          type: string
        difficulty:
          type: string
        duration:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required: [id, title, category, type, difficulty, duration, questions]
    Question:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [mcq, coding, text]
      required: [id, text, type]
    CreateAssessmentBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        type:
          type: string
          enum: [mcq, coding, aptitude, communication]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        duration:
          type: integer
        questions:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              options:
                type: array
                items:
                  type: string
              correctAnswer:
                type: string
              type:
                type: string
            required: [text, type]
      required: [title, category, type, difficulty, duration, questions]
    SubmitAssessmentBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              answer:
                type: string
            required: [questionId, answer]
      required: [answers]
    AssessmentResult:
      type: object
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        userId:
          type: integer
        score:
          type: number
        passed:
          type: boolean
        certificate:
          type: string
        completedAt:
          type: string
        assessment:
          $ref: "#/components/schemas/Assessment"
      required: [id, assessmentId, userId, score, passed, completedAt]

    # LEARNING
...
    CollegeForm:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [internship, hackathon, leave, project, assignment, other]
        description:
          type: string
        deadline:
          type: string
        fields:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              label:
                type: string
              type:
                type: string
              required:
                type: boolean
            required: [name, label, type, required]
        createdById:
          type: integer
        status:
          type: string
          enum: [open, closed]
        submissionCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, type, description, status]
    CreateCollegeFormBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [internship, hackathon, leave, project, assignment, other]
        description:
          type: string
        deadline:
          type: string
        fields:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              label:
                type: string
              type:
                type: string
              required:
                type: boolean
            required: [name, label, type, required]
      required: [title, type, description, fields]
    FormSubmission:
      type: object
      properties:
        id:
          type: integer
        formId:
          type: integer
        userId:
          type: integer
        data:
          type: object
        status:
          type: string
          enum: [pending, approved, rejected]
        feedback:
          type: string
        submittedAt:
          type: string
        form:
          $ref: "#/components/schemas/CollegeForm"
        user:
          $ref: "#/components/schemas/User"
      required: [id, formId, userId, data, status, submittedAt]
    CreateSubmissionBody:
      type: object
      properties:
        formId:
          type: integer
        data:
          type: object
      required: [formId, data]
    UpdateSubmissionBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        feedback:
          type: string
      required: [status]
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [general, event, deadline, hackathon, job]
        createdById:
          type: integer
        createdAt:
          type: string
      required: [id, title, content, type, createdAt]
    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [general, event, deadline, hackathon, job]
      required: [title, content, type]

    # NOTIFICATIONS
...